Size trimming
Implemented a method of shaving off 2 bytes for each of the members
inside of an instance of Object which require access to other members'
data (Behavior, Sprite/Animation, and Input). So, Behavior is now the
same size as an int- just 4 bytes- and is set up to be able to do quite
a bit. I am very happy with this, but am going to have to be adding a
few more bytes for extra functionality later on.

+++ player.h
@@ -7,7 +7,7 @@ class Player : public Object
 {
 	private:
 		//Private member variables
-		int player_Pointer_Vector_Index;			//The pointer's position in the static vector of player pointers
+		uint16_t player_Pointer_Vector_Index;			//The pointer's position in the static vector of player pointers
 
 	public:
 		//Static class variables; currently public, may be made private in the future
